main.js разбит на компоненты
