+++ src/components/trip-list.js
@@ -1,0 +1,59 @@
+import { createForm } from './form.js'
+
+const createTripDays = (dayList) => `
+          <ul class="trip-days">
+          	${createForm()}
+            ${dayList.map((item) => `
+            <li class="trip-days__item  day">
+              <div class="day__info">
+                <span class="day__counter">${item.dayCounter}</span>
+                <time class="day__date" datetime="${item.dayDate}">${item.dayDateText}</time>
+              </div>
+              ${createEventList(item.eventList, item.dayDate)}
+            </li>`).join(`\n`)}
+          </ul>
+`;
+
+const createEventList = (eventList, date) => `
+	              <ul class="trip-events__list">
+	              	${eventList.map((item) => `
+	                <li class="trip-events__item">
+	                  <div class="event">
+	                    <div class="event__type">
+	                      <img class="event__type-icon" width="42" height="42" src="img/icons/${item.type}.png" alt="Event type icon">
+	                    </div>
+	                    <h3 class="event__title">${item.title}</h3>
+
+	                    <div class="event__schedule">
+	                      <p class="event__time">
+	                        <time class="event__start-time" datetime="${date}T${item.startTime}">${item.startTime}</time>
+	                        &mdash;
+	                        <time class="event__end-time" datetime="${date}T${item.endTime}">${item.endTime}</time>
+	                      </p>
+	                      <p class="event__duration">${item.duration}</p>
+	                    </div>
+
+	                    <p class="event__price">
+	                      &euro;&nbsp;<span class="event__price-value">${item.price}</span>
+	                    </p>
+	                    ${item.offers.length ? `
+	                    <h4 class="visually-hidden">Offers:</h4>
+	                    <ul class="event__selected-offers">
+	                    	${item.offers.map((offer) => `
+	                      <li class="event__offer">
+	                        <span class="event__offer-title">${offer.title}</span>
+	                        &plus;
+	                        &euro;&nbsp;<span class="event__offer-price">${offer.price}</span>
+	                       </li>`).join(`\n`)}
+	                    </ul>` : ``}
+
+	                    <button class="event__rollup-btn" type="button">
+	                      <span class="visually-hidden">Open event</span>
+	                    </button>
+	                  </div>
+	                </li>`).join(`\n`)}
+	              </ul>
+`;
+
+
+export { createTripDays }
